backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
cffi==2.0.0
click==8.3.3
cryptography==48.0.0
dnspython==2.8.0
ecdsa==0.19.2
fastapi==0.136.1
h11==0.16.0
idna==3.13
passlib==1.7.4
pyasn1==0.6.3
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pymongo==4.17.0
python-dotenv==1.2.2
python-jose==3.5.0
redis==7.4.0
rsa==4.9.1
six==1.17.0
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
